[POM-200] structure fixes
