+++ src/app/give-help/transport-form/transport-form.component.ts
@@ -3,9 +3,9 @@ import { defaults } from '@app/shared/utils';
 import { TransportOfferDefinitionDTO } from '@app/core/api/model/transportOfferDefinitionDTO';
 import { TransportResourceService } from '../../core/api/api/transportResource.service';
 import { PREFIXES } from '@app/shared/consts';
-import { CorePath } from '@app/shared/models/core-path.enum';
+import { CorePath } from '@app/shared/models/core-path.model';
 import { SnackbarService } from '@app/shared/services/snackbar.service';
-import { ALERT_TYPES } from '@app/shared/models/temporary-models';
+import { ALERT_TYPES } from '@app/shared/models';
 import { take } from 'rxjs/operators';
 import { Router } from '@angular/router';
 // Waiting for TransportOfferDefinitionDTO receive a phoneNumber prop
