+++ src/app/shared/models/core-path.model.ts
