+++ src/app/shared/models/index.ts
@@ -1,6 +1,8 @@
-export * from './category-name-key.enum';
-export * from './category';
-export * from './breadcrumb';
-export * from './breadcrumb-labels';
-export * from './core-path.enum';
-export * from './routing-category-name.enum';
+export * from './breadcrumb-labels.model';
+export * from './breadcrumb.model';
+export * from './category-name-key.model';
+export * from './category.model';
+export * from './core-path.model';
+export * from './offer.model';
+export * from './routing-category-name.model';
+export * from './snackbar.model';
