+++ src/app/shared/models/offer.model.ts
@@ -14,14 +14,3 @@ export interface Option {
   code: string;
   label: string;
 }
-
-export class Alert {
-  header: string | undefined;
-  content: string | undefined;
-}
-
-export enum ALERT_TYPES {
-  CONFIRM = 'CONFIRM',
-  ERROR = 'ERROR',
-  WARNING = 'WARNING',
-}
