+++ src/app/shared/models/snackbar.model.ts
@@ -1,0 +1,10 @@
+export class Alert {
+  header: string | undefined;
+  content: string | undefined;
+}
+
+export enum ALERT_TYPES {
+  CONFIRM = 'CONFIRM',
+  ERROR = 'ERROR',
+  WARNING = 'WARNING',
+}
