+++ src/assets/styles/_snackbar.scss
@@ -1,0 +1,46 @@
+@import "vars";
+
+snack-bar-container {
+  font-size: 15px;
+  &.snack-bar-error {
+    background: $c-lightgray;
+    border: 2px solid $c-redalert;
+    font-weight: 500;
+    color: $c-black;
+  }
+  &.snack-bar-warning {
+    background: $c-lightgray;
+    border: 2px solid $c-yellowwarning;
+    font-weight: 500;
+    color: $c-black;
+  }
+  &.snack-bar-confirm {
+    background: $c-lightgray;
+    border: 2px solid $c-greenconfirm;
+    font-weight: 500;
+    color: $c-black;
+  }
+  &.snackbar-alert {
+    color: $c-black;
+    max-width: 90vw !important;
+    width: 100%;
+    background: $c-lightgray;
+    z-index: 10000;
+    margin: 20vh auto !important;
+    border: 2px solid $c-green;
+    @media (min-width: $max-tab) {
+      margin: 16vh auto !important;
+    }
+    h5 {
+      font-weight: bold;
+    }
+    .button-icon {
+      width: 20%;
+      @media (min-width: $max-tab) {
+        width: 10%;
+      }
+      color: $c-green;
+      font-size: 3em;
+    }
+  }
+}
