+++ src/styles.scss
@@ -4,6 +4,7 @@
 @import "assets/styles/forms";
 @import "assets/styles/utilities";
 @import "assets/styles/overrides";
+@import "assets/styles/snackbar";
 
 html,
 body {
